Added IVaultManagerService
